convert each "size_t iov_len" declaration to "unsigned int iov_len"
Via this command:
git grep -l -E 'size_t[[:blank:]]+iov_len' \
  | xargs perl -pi -e 's/((?:^|,)\s*)size_t\s+(iov_len)/${1}unsigned int $2/'

Affected files and (functions/macros):
* exec/totempg.c (totempg_groups_mcast_groups)
(totempg_groups_send_ok_groups):
* include/corosync/evs.h (evs_callbacks_t):
* include/corosync/totem/totempg.h (TOTEMPG_SAFE):
* lib/evs.c (evs_mcast_joined, evs_mcast_groups):
* man/cpg_mcast_joined.3:
* man/evs_mcast_groups.3:
* man/evs_mcast_joined.3:

git-svn-id: http://svn.fedorahosted.org/svn/corosync/trunk@2032 fd59a12c-fef9-0310-b244-a6a79926bd2f

+++ include/corosync/evs.h
@@ -142,7 +142,7 @@ cs_error_t evs_mcast_joined (
 	evs_handle_t handle,
 	evs_guarantee_t guarantee,
 	const struct iovec *iovec,
-	size_t iov_len);
+	unsigned int iov_len);
 
 /*
  * Multicast to specified groups.
@@ -155,7 +155,7 @@ cs_error_t evs_mcast_groups (
 	const struct evs_group *groups,
 	size_t group_cnt,
 	const struct iovec *iovec,
-	size_t iov_len);
+	unsigned int iov_len);
 
 /*
  * Get membership information from evs
